backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
duckduckgo-search
googlesearch-python
wikipedia
